Updated git submodules
Project: netvirt master 5bbcb2e79d789667112590b1808aa5fc5db4cbe7

Remove .0 from model.bgp.version

The version bump script ancorrectly added a .0 to the version.
This was a problem as during this same time the vpnservice-features
was using the value without the .0 from a different patch, so they
were not consistent.

Change-Id: I38c47898bdf706ccb0f89ce76209cbbbbb8540d8
Signed-off-by: Sam Hague <shague@redhat.com>