Updated git submodules
Project: netvirt stable/carbon e78e0f6d2b0f6d8c33c660e7e9dc7a5b7cf5758e

Bug 8235 - Bug in group programming for directly connected subnet routes

Issue:

1. The ECMP feature enables the loadbalancing of traffic to the the extra
routes only but the discovered routes are also getting loadbalanced similar
to the extra routes.

2. This also leads to the issue in case of the discovered routes
containing macaddress(directly connected subnet route), the
loadbalancing groupId overrides the local nexthop group Id
as the parameters used for computing the groupId turns out to be same.

Analysis:

1. Since the discovered routes are also present in the vpntoextraroute
container, the discovered routes are also being currently loadbalanced by
ECMP feature.

2. Also, the loadbalancing groupId is computed from the prefix or the nexthop Ip
depending on whether the macadress is present or not(if mac is present,
the nexthopIp is used and prefix for otherwise). In case of the directly
connected subnet route, the macaddress is present and so the nexthopIp
is used for computing the loadbalancing groupId. This overlaps with
local next hop group programmed for that nexthop.

Fix:

Since the discovered routes have only 1 next hop, this
condition is now used to distinguish between discovered routes and the
extra routes. Only the extra routes are loadbalanced now and the
discovered routes are handled as before(creating a new local nexthop
group).

Change-Id: I6b0c5506f438ec1f816dfa0fdb045a8494950f48
Signed-off-by: gobinath <gobinath@ericsson.com>