Update patch set 2
Patch Set 2:

> > This is added only on the root level pom. If we need to disable a
 > > test which is run in more than one project, is it require to
 > modify
 > > the poms in the project aswell ?
 > 
 > I guess we'll find out. My guess is that unless it's overridden,
 > things should honor this setting.

This would lead me to believe it will work:
https://maven.apache.org/guides/mini/guide-configuring-plugins.html#Using_the_inherited_Tag_In_Build_Plugins

That being said, the pom files in the projects aren't really children of this root pom.xml file. If the configuration needs to be done in the real parents, then we would need to do more complicated things, e.g., put this into the relevant odlparent pom.

Patch-set: 2