Updated git submodules
Project: alto stable/carbon c1b9623408cc37c8ed92eace62e568219145f5b0

Mark containers as presence

Adding a mandatory node via an augmentation directly into a container
is not allowed, as it changes the semantics. Since this is already
encapsulated, just mark the containers as presence, which makes
the nodes mandatory only if the container is present.

Change-Id: I5755e08e19adf5454b58fdddb4795a057ac4f091
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>