Updated git submodules
Project: netvirt stable/carbon e19a8ac2221c908ac5a658fd16b40638210dd9a5

BUG 8537: Get destination from interface remote ip

Genius will not set the tunnel destination IP address until the very
last action. Classifier egress cannot expect to have this IP address
available on it's own pipeline.

Solve this by setting the IP address on REG0 on the egress service
binding actions. This will be the remote IP of the bound interface in
case of tunnel or the local IP otherwise.

To avoid overuse of registries, a REG is no longer used to set the
SFF IP address on table 101. This IP will be set directly on the path
specific flows of table 223 if needed. Previously this REG was being
used to know if the SFF was remote by comparing with the local IP in
these flows but that is not needed as it can be known at configuration
time.

Change-Id: If7a474391ed2d643cc3b3f34743d60c5f418b2e5
Signed-off-by: Jaime Caamaño Ruiz <jaime.caamano.ruiz@ericsson.com>